refactor scores
